feat: main page pathname 변경
