+++ src/pages/index.tsx
@@ -1,39 +1,93 @@
-import type { CustomNextPage } from 'next/types';
+import type { GetStaticProps } from 'next';
 
-import { useRouter } from 'next/router';
+import { css } from '@emotion/react';
+import { QueryClient } from '@tanstack/react-query';
 
-import { useEffect } from 'react';
-
-import { FullPageLoader, loaderText } from '~/components/Common/FullPageLoader';
-import { useMyAccountStatus, useMyInfo } from '~/services/member';
+import { Clock } from '~/components/Clock';
+import { PageHead } from '~/components/Common/PageHead';
+import { PageHeadingText } from '~/components/Common/PageHeadingText';
+import { HotArticlesPreview } from '~/components/HotArticlesPreview';
+import { LunchMenusPreview } from '~/components/Lunch';
+import NavigationGroup from '~/components/NavigationGroup';
+import { RecruitsPreview } from '~/components/RecruitsPreview';
+import { queryKeys } from '~/react-query/common';
+import { dehydrate } from '~/react-query/server';
+import { toSSRSafeDehydratedState } from '~/react-query/server/toSSRSafeDehydratedState';
+import { getHotArticles } from '~/services/article/apis';
+import { useMyInfo } from '~/services/member';
+import { getRecruits } from '~/services/recruit';
+import { globalVars, topBarHeight } from '~/styles/utils';
 import { routes } from '~/utils';
-import { webStorage } from '~/utils/webStorage';
+import { globalMetaData } from '~/utils/metadata';
 
-const HomePage: CustomNextPage = () => {
-  const router = useRouter();
-  const { isFetching } = useMyInfo({ enabled: true, retry: 1 });
-  const { isRegisterRequired } = useMyAccountStatus();
+const metaTitle = '메인';
+const metaDescription = `${globalMetaData.description} 점심 메뉴, 리쿠르팅, 핫 게시글 등 다양한 SSAF SOUND의 기능을 활용해보세요.`;
 
-  useEffect(() => {
-    if (isFetching) return;
+const MainPage = () => {
+  const { data: myInfo } = useMyInfo();
+  const myCampus = myInfo?.ssafyInfo?.campus;
 
-    // User Register가 필요한 경우에는 <Background />에서 Register 페이지로 리다이렉션을 트리거합니다.
-    if (!isRegisterRequired) {
-      const returnPage = webStorage.getSignInReturnPage();
+  return (
+    <>
+      <PageHead
+        title={metaTitle}
+        description={metaDescription}
+        openGraph={{
+          title: metaTitle,
+          description: metaDescription,
+          url: routes.main(),
+        }}
+      />
 
-      router.replace(returnPage).then(() => {
-        webStorage.clearSignInReturnPage();
-      });
-    }
-  }, [isRegisterRequired, router, isFetching]);
+      <PageHeadingText text={metaTitle} />
 
-  return <FullPageLoader text={loaderText.checkUser} />;
+      <div css={selfCss}>
+        <NavigationGroup />
+        <Clock css={{ marginBottom: 32 }} />
+        <LunchMenusPreview css={{ marginBottom: 80 }} campus={myCampus} />
+        <HotArticlesPreview css={{ marginBottom: 50 }} />
+        <RecruitsPreview
+          css={{ marginBottom: 50 }}
+          marginForExpand={globalVars.mainLayoutPaddingX.var}
+        />
+      </div>
+    </>
+  );
 };
 
-export default HomePage;
-HomePage.meta = {
-  title: '홈',
-  openGraph: {
-    url: routes.root(),
-  },
+export default MainPage;
+
+const selfPaddingY = topBarHeight + 30;
+
+const selfCss = css({
+  padding: `${selfPaddingY}px 0px`,
+});
+
+export const getStaticProps: GetStaticProps = async () => {
+  const queryClient = new QueryClient();
+
+  const recruitsQueryKey = JSON.parse(JSON.stringify(queryKeys.recruit.list()));
+  const hotArticlesQueryKey = JSON.parse(
+    JSON.stringify(queryKeys.articles.hot())
+  );
+
+  await Promise.allSettled([
+    queryClient.prefetchInfiniteQuery({
+      queryKey: hotArticlesQueryKey,
+      queryFn: () => getHotArticles(),
+    }),
+    queryClient.prefetchInfiniteQuery({
+      queryKey: recruitsQueryKey,
+      queryFn: () => getRecruits(),
+    }),
+  ]);
+
+  const { dehydratedState } = dehydrate(queryClient);
+  return {
+    props: {
+      dehydratedState:
+        toSSRSafeDehydratedState.infiniteQueries(dehydratedState),
+    },
+    revalidate: 30,
+  };
 };
